Add GitHub authentication
- Create layout components
- Web: install axios and jwt-decode packages
- Server: install dotenv, axios and @fastify/jwt dependencies
- Mobile: install axios, expo-auth-session, expo-constants, expo-crypto, expo-linking, expo-router, expo-secure-store, react-native-gesture-handler, react-native-safe-area-context, react-native-screens, metro, metro-resolver

+++ server/src/routes/notes.ts
@@ -3,8 +3,15 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function notesRoutes(app: FastifyInstance) {
-  app.get('/notes', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/notes', async (request) => {
     const notes = await prisma.note.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -19,7 +26,7 @@ export async function notesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/notes/:id', async (request) => {
+  app.get('/notes/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -31,6 +38,10 @@ export async function notesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!note.isPublic && note.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return note
   })
@@ -49,14 +60,14 @@ export async function notesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'c27cd0c4-acb1-41ba-9be3-a433fa95bab0',
+        userId: request.user.sub,
       },
     })
 
     return note
   })
 
-  app.put('/notes/:id', async (request) => {
+  app.put('/notes/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +82,17 @@ export async function notesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const note = await prisma.note.update({
+    let note = await prisma.note.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (note.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    note = await prisma.note.update({
       where: {
         id,
       },
@@ -85,12 +106,22 @@ export async function notesRoutes(app: FastifyInstance) {
     return note
   })
 
-  app.delete('/notes/:id', async (request) => {
+  app.delete('/notes/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const note = await prisma.note.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (note.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.note.delete({
       where: {
